+++ packages/guides/Axis.svelte
@@ -7,202 +7,89 @@
   //   the numeric values, placement, etc. for each. This would let the implementer
   //   do whatever they want without having to make the same tedious calculations.
 
-  import { tweened } from "svelte/motion"; // eslint-disable-line import/no-extraneous-dependencies
-
-  import { getContext, setContext } from "svelte"; // eslint-disable-line import/no-extraneous-dependencies
-  import { fade } from "svelte/transition"; // eslint-disable-line import/no-extraneous-dependencies
-
+  import AxisContainer from "./AxisContainer.svelte";
   import AxisLabel from "./AxisLabel.svelte";
   import AxisLine from "./AxisLine.svelte";
   import AxisTick from "./AxisTick.svelte";
 
   export let side; // left, right, top, bottom
-  export let mainScaleName;
-
-  if (!mainScaleName) {
-    if (side === "left" || side === "right") mainScaleName = "yScale";
-    else if (side === "top" || side === "bottom") mainScaleName = "xScale";
-    else mainScaleName = "xScale";
-  }
-
-  export let mainScale = getContext(mainScaleName);
-  export let bodyDimension = getContext(`${side}Plot`);
-  export let rotate = 0;
-
-  let obverse;
-  let minDim;
-  let maxDim;
-
-  if (side === "left") obverse = "right";
-  if (side === "right") obverse = "left";
-  if (side === "top") obverse = "bottom";
-  if (side === "bottom") obverse = "top";
-
-  if (side === "left" || side === "right") {
-    minDim = "bottom";
-    maxDim = "top";
-  }
-
-  if (side === "top" || side === "bottom") {
-    minDim = "left";
-    maxDim = "right";
-  }
-
-  export let obverseDimension = getContext(`${obverse}Plot`);
-
-  // minimumDimension and maximumDimension represents the
-  // max and min dimensions (eg the ends of the AxisLine)
-  export let minimumDimension = getContext(`${minDim}Plot`);
-  export let maximumDimension = getContext(`${maxDim}Plot`);
-
-  // TEXT ANCHOR
-  let defaultTextAnchor;
-  if (side === "left") {
-    defaultTextAnchor = "end";
-    // if (rotate) defaultTextAnchor = 'middle';
-  } else if (side === "right") {
-    defaultTextAnchor = "start";
-    // if (rotate) defaultTextAnchor = 'middle';
-  } else if (side === "top" || side === "bottom") {
-    defaultTextAnchor = "middle";
-    // if (rotate && side === 'top') defaultTextAnchor = 'end';
-    // if (rotate && side === 'bottom') defaultTextAnchor = 'start';
-  }
-
-  export let align = defaultTextAnchor;
-
-  const defaults = getContext("defaults");
-  const buffer = getContext("gp:datagraphic:buffer");
-
-  export let tickDashArray;
-
-  export let tickCount = 5;
-
-  function orderMagnitude(n) {
-    let order = Math.floor(Math.log(n) / Math.LN10 + 0.000000001); // because float math sucks like that
-    return 10 ** order;
-  }
-
-  function symLogTicks(topVal) {
-    // symLogTicks is needed until https://github.com/d3/d3-scale/issues/162 is resolved.
-    const upper = orderMagnitude(topVal);
-    let current = upper;
-    const ticks = [upper];
-    while (current > 1) {
-      current /= 10;
-      ticks.push(current);
-    }
-    ticks.reverse();
-    return ticks;
-  }
-
-  function getDefaultTicks(sc) {
-    if (sc.type === "numeric" || sc.type === "linear" || sc.type === "time") {
-      return sc.ticks(tickCount);
-    }
-    if (sc.type === "log") {
-      return symLogTicks($mainScale.domain()[1]);
-    }
-    return sc.domain().reduce((acc, v, i, source) => {
-      // let's filter to get the right number of ticks.
-      const every = Math.floor(source.length / tickCount);
-      if (i % every === 0) {
-        acc.push(v);
-      }
-      return acc;
-    }, []);
-  }
+  export let lineStyle = side === "left" ? "long" : "short";
 
   export let ticks;
-  // we will need to internally calculate TICKS depending on the passed value
-  // of ticks.
-  let TICKS;
-  $: if (Array.isArray(ticks)) {
-    TICKS = ticks;
-  } else if (typeof ticks === "function") {
-    // if you pass in a function, the function operates
-    // on the xScale accordingly and returns whatever it needs
-    // to be an array
-    TICKS = ticks($mainScale);
-  } else {
-    TICKS = getDefaultTicks($mainScale);
-  }
+  export let tickFormatter;
 
-  // change TICKS to have x, y, x1, x2, y1, y2
-
-  export let tickDirection = side === "right" || side === "bottom" ? 1 : -1;
-
-  export let fadeValues = defaults.fadeParams;
-  export let tickFontSize = defaults.axisTickFontSize;
-
-  export let lineStyle = side === "left" || side === "right" ? "long" : "short";
-  export let tickFormatter =
-    $mainScale.type === "time" ? $mainScale.tickFormat(tickCount) : (t) => t;
-
-  export let showTicks = true;
-  export let showBorder = false;
   export let showLabels = true;
+  export let showTicks = true;
+  export let showLine = false;
 
-  let mainDim = side === "left" || side === "right" ? "x" : "y";
-  let secondaryDim = side === "left" || side === "right" ? "y" : "x";
+  // AxisLabel, AxisLine, AxisTick props.
+  export let color;
+  export let alpha;
+  export let size;
+  export let dashArray;
+  export let fontSize;
 
-  let fontSizeCorrector = side === "bottom" ? tickFontSize : $buffer;
+  export let tickColor = color;
+  export let tickAlpha = alpha;
+  export let tickSize = size;
+  export let tickDashArray = dashArray;
 
-  // axis system context setting.
-  // children like AxisLabel consume these.
+  export let labelColor = color;
+  export let labelAlpha = alpha;
 
-  setContext("mainDim", mainDim);
-  setContext("secondaryDim", secondaryDim);
-  setContext("mainScale", mainScale);
-  setContext("side", side);
-  setContext("bodyDimension", bodyDimension);
-  setContext("obverseDimension", obverseDimension);
-  setContext("minimumDimension", minimumDimension);
-  setContext("maximumDimension", maximumDimension);
-
-  setContext("tickDirection", tickDirection);
-
-  setContext("fontSizeCorrector", fontSizeCorrector);
-  setContext("tickFormatter", tickFormatter);
-  setContext("align", align);
-
-  let R = tweened(-1, { duration: 500 });
-  setInterval(() => {
-    $R = $R < 0 ? 1 : -1;
-  }, 1000);
+  export let lineColor = color;
+  export let lineAlpha = alpha;
+  export let lineSize = size;
+  export let lineDashArray = dashArray;
 </script>
 
-<g in:fade={fadeValues} class="{side}-axis">
+<AxisContainer {tickFormatter} {ticks} {side}>
 
-  {#if showTicks}
-    <slot name="ticks" ticks={TICKS}>
-      {#each TICKS as tick, i (tick)}
-        <AxisTick dashArray={tickDashArray} placement={tick} />
+  <g slot="ticks" let:ticks let:closestMargin let:farthestMargin>
+    {#if showTicks}
+      {#each ticks as tick, i (tick)}
+        <AxisTick
+          color={tickColor}
+          alpha={tickAlpha}
+          dashArray={tickDashArray}
+          size={tickSize}
+          placement={tick} />
         {#if lineStyle === 'long'}
           <AxisTick
             dashArray={tickDashArray}
             placement={tick}
+            size={tickSize}
             tickDirection={-1}
-            length={$bodyDimension - $obverseDimension} />
+            color={tickColor}
+            alpha={tickAlpha}
+            length={closestMargin - farthestMargin} />
         {/if}
       {/each}
-    </slot>
-  {/if}
-
-  <slot name="border">
-    {#if showBorder}
-      <AxisLine />
     {/if}
-  </slot>
+  </g>
 
-  {#if showLabels}
-    <slot name="labels" {tickFormatter} ticks={TICKS}>
-      {#each TICKS as tick, i}
-        <AxisLabel {rotate} placement={tick} {align}>
+  <g slot="border">
+    {#if showLine}
+      <AxisLine
+        color={lineColor}
+        alpha={lineAlpha}
+        size={lineSize}
+        dashArray={lineDashArray} />
+    {/if}
+  </g>
+
+  <g slot="labels" let:ticks let:tickFormatter>
+    {#if showLabels}
+      {#each ticks as tick, i}
+        <AxisLabel
+          {fontSize}
+          color={labelColor}
+          alpha={labelAlpha}
+          placement={tick}>
           {tickFormatter ? tickFormatter(tick) : tick}
         </AxisLabel>
       {/each}
-    </slot>
-  {/if}
+    {/if}
+  </g>
 
-</g>
+</AxisContainer>