+++ packages/guides/AxisLabel.svelte
@@ -1,40 +1,36 @@
 <script>
-  import { getContext, onMount } from "svelte"; // eslint-disable-line import/no-extraneous-dependencies
+  import { getContext } from "svelte"; // eslint-disable-line import/no-extraneous-dependencies
 
-  export let side = getContext("side");
-  export let mainDim = getContext("mainDim");
-  export let secondaryDim = getContext("secondaryDim");
-  export let mainScale = getContext("mainScale");
-  export let bodyDimension = getContext("bodyDimension");
-  export let tickDirection = getContext("tickDirection");
-  export let fontSizeCorrector = getContext("fontSizeCorrector");
   export let buffer = getContext("gp:datagraphic:buffer");
-  export let tickFormatter =
-    getContext("tickFormatter") ||
-    function format(v) {
-      return v;
-    };
-  export let align = getContext("align") || "middle";
 
+  export let side = getContext("gp:axis:side");
+  export let response = getContext("gp:axis:response");
+  export let orientation = getContext("gp:axis:orientation");
+  export let scale = getContext("gp:axis:scale");
+  export let closestMargin = getContext("gp:axis:closestMargin");
+  export let tickDirection = getContext("gp:axis:tickDirection");
+  export let tickFormatterStore = getContext("gp:axis:tickFormatterStore");
+  export let tickFormatter;
+
+  export let color = "var(--cool-gray-700)";
+  export let alpha = 1;
   // the domain value where the placement should occur
   export let placement;
-  export let offset = 0;
-  export let rotate = 0;
-  export let fontSize = 10;
+  export let fontSize = "10px";
   export let fontWeight = "normal";
-  export let color = "black";
-  export let dx = 0;
-  export let dy = 0;
 
-  function place(v, dim, sc, buff) {
-    if (mainDim === dim) {
+  $: if (!tickFormatter) {
+    tickFormatter = $tickFormatterStore;
+  } else {
+    tickFormatter = (v) => v;
+  }
+
+  function place(v, dim, sc, bd, buff) {
+    if (response === dim) {
       return (
-        $bodyDimension +
+        bd +
         tickDirection * buff +
-        (side === "top" || side === "bottom"
-          ? tickDirection * fontSizeCorrector
-          : 0) +
-        (side === "left" || side === "top" ? -offset : offset)
+        (side === "left" || side === "top" ? -buff : buff)
       );
     }
     let step = sc.type === "scaleBand" ? sc.bandwidth() / 2 : 0;
@@ -42,42 +38,44 @@
   }
 
   // let's set alignment dynamically as a dx property
-  let container;
-  let label;
-  let alignmentOffset = 0;
-  let yAdjustment = 0;
-  let mounted = false;
-  onMount(() => {
-    mounted = true;
-  });
 
-  $: if (mounted && align) {
-    let { width, height } = container.getBoundingClientRect();
-    // let { height } = label.getBoundingClientRect();
-    if (align === "middle") alignmentOffset = -width / 2;
-    else if (align === "end") alignmentOffset = -width;
-    else alignmentOffset = 0;
-    yAdjustment = height / 4;
-  }
+  let textAnchor = "middle";
+  if (side === "left") textAnchor = "end";
+  if (side === "right") textAnchor = "start";
 
-  let transform = "";
-  $: transform = `${
-    rotate !== 0
-      ? `rotate(${rotate} ${
-          place(placement, "x", $mainScale, $buffer) - alignmentOffset
-        } ${place(placement, "y", $mainScale, $buffer)}) `
-      : " "
-  } ${mainDim === "x" ? `translate(0 ${yAdjustment})` : ""}`;
+  let dy;
+  $: if (orientation === "y") {
+    dy = ".35em";
+  } else if (side === "bottom") {
+    dy = "1em";
+  } else dy = undefined;
+
+  $: parameters = {
+    [`${response}`]: place(
+      placement,
+      response,
+      $scale,
+      $closestMargin,
+      $buffer
+    ),
+    [`${orientation}`]: place(
+      placement,
+      orientation,
+      $scale,
+      $closestMargin,
+      $buffer
+    ),
+    dy,
+    "text-anchor": textAnchor,
+  };
 </script>
 
-<g bind:this={container} transform="translate({alignmentOffset} 0)">
+<g>
   <text
-    bind:this={label}
-    {...{ [`${mainDim}`]: place(placement, mainDim, $mainScale, $buffer), [`${secondaryDim}`]: place(placement, secondaryDim, $mainScale, $buffer), dy, dx }}
-    font-size={fontSize}
-    font-weight={fontWeight}
+    {...parameters}
+    style="font-size: {fontSize}; font-weight: {fontWeight};"
     fill={color}
-    {transform}>
+    opacity={alpha}>
     <slot>{tickFormatter(placement)}</slot>
   </text>
 </g>